+++ src/main.js
@@ -4,6 +4,7 @@ import normalizeException from 'normalize-exception'
 import { normalizeOptions, validateOptions } from './options.js'
 import { parseDeep, parseShallow } from './parse/main.js'
 import { serializeDeep, serializeShallow } from './serialize.js'
+import { applyTransformInstance } from './transform.js'
 
 export { validateOptions }
 
@@ -15,14 +16,15 @@ export { validateOptions }
 // We apply `normalize-exception` to ensure a strict input.
 //  - We allow arguments that are not `error` instances
 export const serialize = (value, options) => {
-  const { loose, shallow, beforeSerialize, afterSerialize } =
-    normalizeOptions(options)
-  const valueA = applyLoose(value, loose)
-  const events = { beforeSerialize, afterSerialize }
+  const { loose, shallow, transformObject } = normalizeOptions(options)
+  const valueA = applySerializeLoose(value, loose)
   return shallow
-    ? serializeShallow(valueA, events)
-    : serializeDeep(valueA, events, [])
+    ? serializeShallow(valueA, transformObject)
+    : serializeDeep(valueA, transformObject, [])
 }
+
+const applySerializeLoose = (value, loose) =>
+  loose || isErrorInstance(value) ? value : normalizeException(value)
 
 // Normalize and convert an already parsed plain object representing an error
 // into an error instance.
@@ -31,14 +33,20 @@ export const serialize = (value, options) => {
 //    some error handling logic
 // We apply `normalize-exception` to ensure a strict output.
 export const parse = (value, options) => {
-  const { loose, shallow, beforeParse, afterParse, classes } =
+  const { loose, shallow, transformArgs, transformInstance, classes } =
     normalizeOptions(options)
-  const events = { beforeParse, afterParse }
   const valueA = shallow
-    ? parseShallow(value, events, classes)
-    : parseDeep(value, events, classes)
-  return applyLoose(valueA, loose)
+    ? parseShallow({ value, transformArgs, transformInstance, classes })
+    : parseDeep({ value, transformArgs, transformInstance, classes })
+  return applyParseLoose(valueA, loose, transformInstance)
 }
 
-const applyLoose = (value, loose) =>
-  loose || isErrorInstance(value) ? value : normalizeException(value)
+const applyParseLoose = (value, loose, transformInstance) => {
+  if (loose || isErrorInstance(value)) {
+    return value
+  }
+
+  const error = normalizeException(value)
+  applyTransformInstance(transformInstance, error, value)
+  return error
+}