+++ src/parse/main.js
@@ -2,27 +2,46 @@ import isPlainObj from 'is-plain-obj'
 import normalizeException from 'normalize-exception'
 
 import { isErrorObject, safeListKeys } from '../check.js'
-import { callEvent } from '../event.js'
 import { listProps, SET_CORE_PROPS, NON_ENUMERABLE_PROPS } from '../props.js'
+import { applyTransformInstance } from '../transform.js'
 
 import { createError } from './create.js'
 
 // Parse error plain objects into error instances deeply
-export const parseDeep = (value, events, classes) => {
-  const valueA = parseRecurse(value, events, classes)
-  return parseShallow(valueA, events, classes)
+export const parseDeep = ({
+  value,
+  transformArgs,
+  transformInstance,
+  classes,
+}) => {
+  const valueA = parseRecurse({
+    value,
+    transformArgs,
+    transformInstance,
+    classes,
+  })
+  return parseShallow({
+    value: valueA,
+    transformArgs,
+    transformInstance,
+    classes,
+  })
 }
 
 // Parse a possible error plain object into an error instance
-export const parseShallow = (value, { beforeParse, afterParse }, classes) => {
+export const parseShallow = ({
+  value,
+  transformArgs,
+  transformInstance,
+  classes,
+}) => {
   if (!isErrorObject(value)) {
     return value
   }
 
-  callEvent(beforeParse, value)
-  const error = parseErrorObject(value, classes)
+  const error = parseErrorObject(value, transformArgs, classes)
   const errorA = normalizeException(error)
-  callEvent(afterParse, value, errorA)
+  applyTransformInstance(transformInstance, errorA, value)
   return errorA
 }
 
@@ -30,8 +49,8 @@ export const parseShallow = (value, { beforeParse, afterParse }, classes) => {
 //  - It does not reuse `normalize-exception`'s object parsing logic
 //  - reason: keep projects separate since they have different purposes and
 //    features
-const parseErrorObject = (errorObject, classes) => {
-  const error = createError(errorObject, classes)
+const parseErrorObject = (errorObject, transformArgs, classes) => {
+  const error = createError({ errorObject, transformArgs, classes })
   setProps(error, errorObject)
   return error
 }
@@ -63,16 +82,23 @@ const setProp = (error, object, propName) => {
   })
 }
 
-const parseRecurse = (value, events, classes) => {
+const parseRecurse = ({ value, transformArgs, transformInstance, classes }) => {
   if (Array.isArray(value)) {
-    return value.map((child) => parseDeep(child, events, classes))
+    return value.map((child) =>
+      parseDeep({ value: child, transformArgs, transformInstance, classes }),
+    )
   }
 
   if (isPlainObj(value)) {
     return Object.fromEntries(
       safeListKeys(value).map((propName) => [
         propName,
-        parseDeep(value[propName], events, classes),
+        parseDeep({
+          value: value[propName],
+          transformArgs,
+          transformInstance,
+          classes,
+        }),
       ]),
     )
   }