+++ src/parse/transform_args.js
@@ -1,0 +1,11 @@
+// Apply `transformArgs` option
+export const applyTransformArgs = ({
+  transformArgs,
+  constructorArgs,
+  errorObject,
+  ErrorClass,
+}) => {
+  try {
+    transformArgs?.(constructorArgs, errorObject, ErrorClass)
+  } catch {}
+}